+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmEditor/AlgorithmEditor.tsx
@@ -1,16 +1,145 @@
 import * as S from "./AlgorithmEditor.styles";
-import CodeMirrorComponent from "./[CY]AlgorithmEditor";
-import CodeMirror from "@uiw/react-codemirror";
 import { javascript } from "@codemirror/lang-javascript";
 import { python } from "@codemirror/lang-python";
 import { java } from "@codemirror/lang-java";
-import { githubDark, githubLight } from "@uiw/codemirror-theme-github";
+import { useEffect, useRef } from "react";
+import yorkie from "yorkie-js-sdk";
+import { basicSetup } from "codemirror";
+import { EditorView, keymap } from "@codemirror/view";
+import { defaultKeymap } from "@codemirror/commands";
+import { Transaction } from "@codemirror/state";
+import "./style.css";
+import { EditorProps } from "./AlgorithmEditor.types";
 
-export const AlgorithmEditor = () => {
+export const AlgorithmEditor = ({ language, isinit }: EditorProps) => {
+  const editorParentRef = useRef();
+  const docRef = useRef();
+  const viewRef = useRef();
+
+  useEffect(() => {
+    console.log("useEffect is running");
+    const initYorkie = async () => {
+      // 01. create client with RPCAddr(envoy) then activate it.
+      const client = new yorkie.Client("https://api.yorkie.dev", {
+        apiKey: "clskuqbj2k70uv115dv0",
+      });
+      await client.activate();
+
+      // 02-1. create a document then attach it into the client.
+      const doc = new yorkie.Document("editor");
+      await client.attach(doc);
+
+      doc.update((root) => {
+        if (!root.content) {
+          root.content = new yorkie.Text();
+        }
+      }, "create content if not exists");
+
+      // 02-2. subscribe document event.
+      const syncText = () => {
+        const text = doc.getRoot().content;
+        viewRef.current.dispatch({
+          changes: {
+            from: 0,
+            to: viewRef.current.state.doc.length,
+            insert: text.toString(),
+          },
+          annotations: [Transaction.remote.of(true)],
+        });
+      };
+
+      doc.subscribe((event) => {
+        if (event.type === "snapshot") {
+          // The text is replaced to snapshot and must be re-synced.
+          syncText();
+        }
+      });
+
+      doc.subscribe("$.content", (event) => {
+        if (event.type === "remote-change") {
+          const { operations } = event.value;
+          handleOperations(operations);
+        }
+      });
+
+      await client.sync();
+
+      // 03-1. define function that bind the document with the codemirror(broadcast local changes to peers)
+      const updateListener = EditorView.updateListener.of((viewUpdate) => {
+        if (viewUpdate.docChanged) {
+          for (const tr of viewUpdate.transactions) {
+            const events = [
+              "select",
+              "input",
+              "delete",
+              "move",
+              "undo",
+              "redo",
+            ];
+            if (!events.map((event) => tr.isUserEvent(event)).some(Boolean)) {
+              continue;
+            }
+            if (tr.annotation(Transaction.remote)) {
+              continue;
+            }
+            tr.changes.iterChanges((fromA, toA, _, __, inserted) => {
+              doc.update((root) => {
+                root.content.edit(fromA, toA, inserted.toJSON().join("\n"));
+              }, `update content byA ${client.getID()}`);
+            });
+          }
+        }
+      });
+
+      // 03-2. create codemirror instance
+      const view = new EditorView({
+        doc: "",
+        extensions: [
+          keymap.of(defaultKeymap),
+          basicSetup,
+          updateListener,
+          python(),
+        ],
+        parent: editorParentRef.current,
+      });
+
+      // Save references to state
+      docRef.current = doc;
+      viewRef.current = view;
+
+      // 03-3. define event handler that apply remote changes to local
+      function handleOperations(operations) {
+        operations.forEach((op) => {
+          if (op.type === "edit") {
+            handleEditOp(op);
+          }
+        });
+      }
+
+      function handleEditOp(op) {
+        const changes = [
+          {
+            from: Math.max(0, op.from),
+            to: Math.max(0, op.to),
+            insert: op.value.content,
+          },
+        ];
+
+        viewRef.current.dispatch({
+          changes,
+          annotations: [Transaction.remote.of(true)],
+        });
+      }
+
+      syncText();
+    };
+
+    initYorkie();
+  }, []);
+
   return (
     <S.AlgorithmEditorContainer>
-      {/* <CodeMirror theme={githubLight} extensions={[python()]} height="64.9vh" /> */}
-      <CodeMirrorComponent />
+      <div ref={editorParentRef} className="editor" />
     </S.AlgorithmEditorContainer>
   );
 };
