+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmEditor/AlgorithmEditor.types.ts
@@ -1,0 +1,4 @@
+export interface EditorProps {
+  language: string;
+  isinit: boolean;
+}
