+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmProHeader/AlgorithmProHeader.types.ts
@@ -1,0 +1,3 @@
+export interface LanguageProps {
+  setLanguage: React.Dispatch<React.SetStateAction<string>>;
+}
